deploy8

+++ database/migrations/2022_03_02_112458_create_carritos_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('carritos', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('id_producto');
+            $table->unsignedBigInteger('id_producto');
             $table->foreign('id_producto')->references('id')->on('productos');
-            $table->unsignedInteger('id_pedido');
+            $table->unsignedBigInteger('id_pedido');
             $table->foreign('id_pedido')->references('id')->on('pedidos');
             $table->timestamps();
         });
